requirements.txt
dgl @ https://data.dgl.ai/wheels/cu121/dgl-0.8.1-cp39-cp39-manylinux1_x86_64.whl
torch @ https://download.pytorch.org/whl/cu121/torch-1.10.2%2Bcu111-cp39-cp39-linux_x86_64.whl
torchvision @ https://download.pytorch.org/whl/cu121/torchvision-0.11.3%2Bcu111-cp39-cp39-linux_x86_64.whl
torchaudio @ https://download.pytorch.org/whl/cu121/torchaudio-0.10.2%2Bcu111-cp39-cp39-linux_x86_64.whl
pydance
